decoupled websockets-extension from core now indpendent from filter-ext, session-exclude-settings, database-creation, etc. introduced new PersistentConnectionListener to keep TCP connection open added possibility to change icon of SiteMap-Node via SiteMapListener.onReturnNodeRendererComponent(...) prepared core for 'Server-Sent Events' extension

+++ src/org/zaproxy/zap/extension/websocket/WebSocketProxyV13.java
@@ -354,7 +354,7 @@ public class WebSocketProxyV13 extends WebSocketProxy {
 		private String getByteAsBitString(byte word) {
 	      StringBuilder buf = new StringBuilder();
 	      for (int i = 0; i < 8; i++) {
-	         buf.append((int)(word >> (8 - (i+1)) & 0x0001));
+	         buf.append((word >> (8 - (i+1)) & 0x0001));
 	      }
 	      return buf.toString();
 	   }
